perf(mantle): decouple cross-component helper imports
Importing a helper from another component's implementation module pulls
that whole module — and its dependencies — into the importer's graph:

- dialog/primitive.tsx imported preventCloseOnPromptInteraction from
  toast/toast.tsx, dragging sonner (which injects CSS at import time and
  declares no sideEffects flag, so bundlers can never drop it) into every
  dialog-only consumer bundle. Moved to its own module; a dialog-only
  Vite 8 build shrinks 117.0 KB -> 86.0 KB (-26.5%).
- calendar/calendar.tsx imported iconButtonVariants from icon-button.tsx
  without using the component. Moved the variants to their own module.

All other cross-component imports were audited: they either compose real
components or import from already-standalone helper modules
(field/validation, field/field-context, theme/themes, input/is-input).
