+++ packages/mantle/src/components/toast/toast.tsx
@@ -461,35 +461,6 @@ export type {
 	ToastIntent,
 };
 
-/**
- * @private
- *
- * Allows any mantle floating prompt (e.g. toasts and notifications) to be interacted with
- * even when a modaled view (e.g. dialog, sheet, etc) is open and a focus trap is active.
- *
- * Without this, interacting with the prompt would close the modaled view.
- *
- * @example
- * ```tsx
- * <Dialog.Root onInteractOutside={preventCloseOnPromptInteraction}>
- *   <Dialog.Content>
- *     <p>Dialog content</p>
- *   </Dialog.Content>
- * </Dialog.Root>
- * ```
- */
-export function preventCloseOnPromptInteraction(
-	event: CustomEvent | PointerEvent | MouseEvent | TouchEvent | FocusEvent,
-) {
-	if (!(event.target instanceof Element)) {
-		return;
-	}
-
-	if (event.target.closest(".overlay-prompt")) {
-		event.preventDefault();
-	}
-}
-
 const intentBackgroundColor = {
 	info: "bg-accent-600",
 	warning: "bg-warning-600",
